New screenplay editor, pt. 1
Implemented most of features. Subtitle clip generation, specifying
color codes for voices and cleaning up the layout is all left

+++ engine/src/main/java/mchorse/bbs/audio/Wave.java
@@ -49,7 +49,7 @@ public class Wave
 
     public float getDuration()
     {
-        return this.data.length / this.numChannels / this.getBytesPerSample() / (float) this.sampleRate;
+        return this.data.length / (float) this.numChannels / (float) this.getBytesPerSample() / (float) this.sampleRate;
     }
 
     public int getALFormat()
